Search history, thumbnail/preview changes, fixes

+++ lib/libBooru/DBHandler.dart
@@ -22,7 +22,7 @@ class DBHandler{
     await deleteUntracked();
   }
   Future<bool> updateTable() async{
-    await db!.execute("CREATE TABLE IF NOT EXISTS BooruItem"
+    await db?.execute("CREATE TABLE IF NOT EXISTS BooruItem"
         "(id INTEGER PRIMARY KEY,"
         "thumbnailURL TEXT,"
         "sampleURL TEXT,"
@@ -32,13 +32,20 @@ class DBHandler{
         "isSnatched INTEGER,"
         "isFavourite INTEGER"
         ")");
-    await db!.execute("CREATE TABLE IF NOT EXISTS Tag ("
+    await db?.execute("CREATE TABLE IF NOT EXISTS Tag ("
         "id INTEGER PRIMARY KEY,"
         "name TEXT"
         ")");
-    await db!.execute("CREATE TABLE IF NOT EXISTS ImageTag ("
+    await db?.execute("CREATE TABLE IF NOT EXISTS ImageTag ("
         "tagID INTEGER,"
         "booruItemID INTEGER"
+        ")");
+    await db?.execute("CREATE TABLE IF NOT EXISTS SearchHistory ("
+        "id INTEGER PRIMARY KEY,"
+        "booruType TEXT,"
+        "booruName TEXT,"
+        "searchText TEXT,"
+        "timestamp TEXT DEFAULT CURRENT_TIMESTAMP"
         ")");
     return true;
   }
@@ -46,30 +53,30 @@ class DBHandler{
   //Inserts a new booruItem or updates the isSnatched and isFavourite values of an existing BooruItem in the database
   void updateBooruItem(BooruItem item) async{
     print("updateBooruItem called fileURL is:" + item.fileURL);
-    String itemID = await getItemID(item.fileURL);
-    if (itemID.isEmpty){
-      var result = await db!.rawInsert("INSERT INTO BooruItem(thumbnailURL,sampleURL,fileURL,postURL,mediaType,isSnatched,isFavourite) VALUES(?,?,?,?,?,?,?)",
-          [item.thumbnailURL,item.sampleURL,item.fileURL,item.postURL,item.mediaType,Tools.boolToInt(item.isSnatched),Tools.boolToInt(item.isFavourite)]);
-      itemID = result.toString();
+    String? itemID = await getItemID(item.fileURL);
+    if (itemID == null || itemID.isEmpty) {
+      var result = await db?.rawInsert("INSERT INTO BooruItem(thumbnailURL,sampleURL,fileURL,postURL,mediaType,isSnatched,isFavourite) VALUES(?,?,?,?,?,?,?)",
+          [item.thumbnailURL, item.sampleURL, item.fileURL, item.postURL, item.mediaType, Tools.boolToInt(item.isSnatched), Tools.boolToInt(item.isFavourite)]);
+      itemID = result?.toString();
       updateTags(item.tagsList!, itemID);
     } else {
-      await db!.rawUpdate("UPDATE BooruItem SET isSnatched = ?, isFavourite = ? WHERE id = ?",[Tools.boolToInt(item.isSnatched),Tools.boolToInt(item.isFavourite),itemID]);
+      await db?.rawUpdate("UPDATE BooruItem SET isSnatched = ?, isFavourite = ? WHERE id = ?", [Tools.boolToInt(item.isSnatched), Tools.boolToInt(item.isFavourite), itemID]);
     }
   }
 
 
   //Gets a BooruItem id from the database based on a fileurl
-  Future<String> getItemID(String fileURL) async{
+  Future<String?> getItemID(String fileURL) async{
     var result;
     if (fileURL.contains("s.sankakucomplex.com")){
-      result = await db!.rawQuery("SELECT id FROM BooruItem WHERE fileURL LIKE (?)",[fileURL.split("?")[0]+"%"]);
+      result = await db?.rawQuery("SELECT id FROM BooruItem WHERE fileURL LIKE (?)", [fileURL.split("?")[0]+"%"]);
     } else {
-      result = await db!.rawQuery("SELECT id FROM BooruItem WHERE fileURL IN (?)",[fileURL]);
+      result = await db?.rawQuery("SELECT id FROM BooruItem WHERE fileURL IN (?)", [fileURL]);
     }
-    if (result.isNotEmpty){
+    if (result != null && result.isNotEmpty){
       return result.first["id"].toString();
     } else {
-      return "";
+      return null;
     }
   }
 
@@ -85,17 +92,17 @@ class DBHandler{
       for (int i = 1; i < tags.length; i++){
         questionMarks += ",?";
       }
-      result = await db!.rawQuery(
+      result = await db?.rawQuery(
           "SELECT booruItemID as id FROM ImageTag INNER JOIN Tag on ImageTag.tagID = Tag.id "
               "WHERE Tag.name IN ($questionMarks) GROUP BY booruItemID "
-              "HAVING COUNT(*) = ${tags.length} ORDER BY id DESC LIMIT $limit OFFSET $offset",tags);
+              "HAVING COUNT(*) = ${tags.length} ORDER BY id DESC LIMIT $limit OFFSET $offset", tags);
     } else {
-      result = await db!.rawQuery(
+      result = await db?.rawQuery(
           "SELECT id FROM BooruItem ORDER BY id DESC LIMIT $limit OFFSET $offset");
     }
     print("got results from db");
     print(result);
-    if (result.isNotEmpty){
+    if (result != null && result.isNotEmpty){
       for(int i=0; i < result.length; i++){
         BooruItem? temp = await getBooruItem(result[i]["id"]);
         if (temp != null){
@@ -108,13 +115,13 @@ class DBHandler{
 
   //Creates a BooruItem from a BooruItem id
   Future<BooruItem?> getBooruItem(int itemID) async{
-    var metaData = await db!.rawQuery("SELECT * FROM BooruItem WHERE ID IN (?) AND isFavourite = 1",[itemID]);
+    var metaData = await db?.rawQuery("SELECT * FROM BooruItem WHERE ID IN (?) AND isFavourite = 1",[itemID]);
     BooruItem item;
     List<String> tagsList = [];
-    if (metaData.isNotEmpty){
+    if (metaData != null && metaData.isNotEmpty){
       item = new BooruItem(metaData.first["fileURL"].toString(), metaData.first["fileURL"].toString(), metaData.first["thumbnailURL"].toString(), null,  metaData.first["postURL"].toString(), metaData.first["fileURL"].toString().substring(metaData.first["fileURL"].toString().lastIndexOf(".") + 1));
-      var tags = await db!.rawQuery("SELECT name FROM ImageTag INNER JOIN Tag on ImageTag.tagID = Tag.ID WHERE booruItemID in (?)",[itemID]);
-      tags.forEach((tag) {tagsList.add(tag["name"].toString());});
+      var tags = await db?.rawQuery("SELECT name FROM ImageTag INNER JOIN Tag on ImageTag.tagID = Tag.ID WHERE booruItemID in (?)", [itemID]);
+      tags?.forEach((tag) {tagsList.add(tag["name"].toString());});
       item.isSnatched = Tools.intToBool(int.parse(metaData.first["isSnatched"].toString()));
       item.isFavourite = Tools.intToBool(int.parse(metaData.first["isFavourite"].toString()));
       item.tagsList = tagsList;
@@ -124,28 +131,29 @@ class DBHandler{
     }
   }
   void clearSnatched() async{
-    await db!.rawUpdate("UPDATE BooruItem SET isSnatched = 0");
+    await db?.rawUpdate("UPDATE BooruItem SET isSnatched = 0");
   }
   void clearFavourites() async{
-    await db!.rawUpdate("UPDATE BooruItem SET isFavourite = 0");
+    await db?.rawUpdate("UPDATE BooruItem SET isFavourite = 0");
   }
   //Adds tags for a BooruItem to the database
-  void updateTags(List tags, String itemID) async{
-    String id = "";
+  void updateTags(List tags, String? itemID) async{
+    if(itemID == null) return;
+    String? id = "";
     tags.forEach((tag) async{
       id = await getTagID(tag);
-      if (id.isEmpty){
-        var result = await db!.rawInsert("INSERT INTO Tag(name) VALUES(?)",[tag]);
-        id = result.toString();
+      if (id != null && id!.isEmpty){
+        var result = await db?.rawInsert("INSERT INTO Tag(name) VALUES(?)", [tag]);
+        id = result?.toString();
       }
-      await db!.rawInsert("INSERT INTO ImageTag(tagID, booruItemID) VALUES(?,?)",[id,itemID]);
+      await db?.rawInsert("INSERT INTO ImageTag(tagID, booruItemID) VALUES(?,?)", [id,itemID]);
     });
   }
 
   //Gets a tag id from the database
   Future<String> getTagID(String tagName) async{
-    var result = await db!.rawQuery("SELECT id FROM Tag WHERE name IN (?)",[tagName]);
-    if (result.isNotEmpty){
+    var result = await db?.rawQuery("SELECT id FROM Tag WHERE name IN (?)", [tagName]);
+    if (result != null && result.isNotEmpty){
       return result.first["id"].toString();
     } else {
       return "";
@@ -154,8 +162,8 @@ class DBHandler{
   //Get a list of tags from the database based on an input
   Future<List<String>> getTags(String queryStr, int limit) async{
     List<String> tags = [];
-    var result = await db!.rawQuery("SELECT name FROM Tag WHERE name LIKE (?) LIMIT $limit",["$queryStr%"]);
-    if (result.isNotEmpty){
+    var result = await db?.rawQuery("SELECT name FROM Tag WHERE name LIKE (?) LIMIT $limit",["$queryStr%"]);
+    if (result != null && result.isNotEmpty){
       for (int i = 0; i < result.length; i++){
         tags.add(result[i]["name"].toString());
       }
@@ -163,17 +171,41 @@ class DBHandler{
     return tags;
   }
 
+  // Remove duplicates and add every new search to history table
+  void updateSearchHistory(String searchText, String? booruType, String? booruName) async{
+    await db?.rawDelete("DELETE FROM SearchHistory WHERE searchText=? AND booruType=? AND booruName=?", [searchText, booruType, booruName]);
+    await db?.rawInsert("INSERT INTO SearchHistory(searchText, booruType, booruName) VALUES(?,?,?)", [searchText, booruType, booruName]);
+  }
+
+  Future<List<List<String>>> getSearchHistory() async{
+    var metaData = await db?.rawQuery("SELECT id, searchText, booruType, booruName FROM SearchHistory GROUP BY searchText, booruName ORDER BY id DESC LIMIT 100");
+    List<List<String>> result = [];
+    metaData?.forEach((s) {
+      result.add([s['id'].toString(), s['searchText'].toString(), s['booruType'].toString(), s['booruName'].toString()]);
+    });
+    return result;
+  }
+
+  Future<void> deleteFromSearchHistory(String? id) async{
+    if(id != null) {
+      await db?.rawDelete("DELETE FROM SearchHistory WHERE id IN (?)", [id]);
+    } else {
+      await db?.rawDelete("DELETE FROM SearchHistory WHERE id IS NOT NULL");
+    }
+    return;
+  }
+
   //Return a list of boolean for isSnatched and isFavourite
   Future<List<bool>> getTrackedValues(String fileURL) async{
     List<bool> values = [false,false];
     var result;
     if (fileURL.contains("s.sankakucomplex.com")){
-      result = await db!.rawQuery("SELECT isFavourite,isSnatched FROM BooruItem WHERE fileURL LIKE (?)",[fileURL.split("?")[0]+"%"]);
+      result = await db?.rawQuery("SELECT isFavourite,isSnatched FROM BooruItem WHERE fileURL LIKE (?)", [fileURL.split("?")[0]+"%"]);
     } else {
-      result = await db!.rawQuery("SELECT isFavourite,isSnatched FROM BooruItem WHERE fileURL IN (?)",[fileURL]);
+      result = await db?.rawQuery("SELECT isFavourite,isSnatched FROM BooruItem WHERE fileURL IN (?)", [fileURL]);
     }
 
-    if (result.isNotEmpty){
+    if (result != null && result.isNotEmpty){
       print("file url is: $fileURL");
       print(result.toString());
       values[0] = Tools.intToBool(result.first["isSnatched"]);
@@ -183,8 +215,8 @@ class DBHandler{
   }
   // Deletes booruItems which are no longer favourited or snatched
   Future<bool> deleteUntracked() async{
-    var result = await db!.rawQuery("SELECT id FROM BooruItem WHERE isFavourite = 0 and isSnatched = 0");
-    if (result.isNotEmpty){
+    var result = await db?.rawQuery("SELECT id FROM BooruItem WHERE isFavourite = 0 and isSnatched = 0");
+    if (result != null && result.isNotEmpty){
       result.forEach((element) async{
         deleteItem(int.parse(element["id"].toString()));
       });
@@ -195,7 +227,7 @@ class DBHandler{
   //Deletes a BooruItem and its tags from the database
   void deleteItem(int itemID) async{
     print("DBHandler deleting: $itemID");
-    await db!.rawDelete("DELETE FROM BooruItem WHERE id IN (?)",[itemID]);
-    await db!.rawDelete("DELETE FROM ImageTag WHERE booruItemID IN (?)",[itemID]);
+    await db?.rawDelete("DELETE FROM BooruItem WHERE id IN (?)", [itemID]);
+    await db?.rawDelete("DELETE FROM ImageTag WHERE booruItemID IN (?)", [itemID]);
   }
 }
